Removing superfluous apps and consolidating everything into mbu app.
