--- course/views.py
@@ -1,6 +1,0 @@
-from django.shortcuts import render
-from django.core.context_processors import csrf
-from django.forms import ModelForm
-from course.models import Course, Session
-
-# Create your views here.
