+++ mbu/templates/mbu/view_classes.html
