--- registration/templates/registration/add_council_modal.html
@@ -1,20 +1,0 @@
-<div class="modal fade" id="council_modal" tabindex="-1" role="dialog" aria-labelledby="myModalLabel" aria-hidden="true">
-    <div class="modal-dialog">
-        <div class="modal-content">
-            <div class="modal-header">
-                <button type="button" class="close" data-dismiss="modal" aria-label="Close"><span aria-hidden="true">&times;</span></button>
-                <h4 class="modal-title" id="myModalLabel">{{ council_modal_title }}</h4>
-            </div>
-            <div class="modal-body">
-                <form method="POST" action="/register/council/" id="council_modal_form">
-                    {% csrf_token %}
-                    {{ council_modal_form.as_p }}
-                </form>
-            </div>
-            <div class="modal-footer">
-                <button type="button" class="btn btn-default" data-dismiss="modal">Close</button>
-                <button id="council_submit_button" type="button" class="btn btn-primary" data-dismiss="modal">{{ council_modal_title }}</button>
-            </div>
-        </div>
-    </div>
-</div>
