--- registration/templates/registration/add_troop_modal.html
@@ -1,27 +1,0 @@
-<button type="button" class="btn btn-primary btn-lg" data-toggle="modal" data-target="#troop_modal">
-  Troop not listed? Add it
-</button>
-
-<div class="modal fade" id="troop_modal" tabindex="-1" role="dialog" aria-labelledby="myModalLabel" aria-hidden="true">
-    <div class="modal-dialog">
-        <div class="modal-content">
-            <div class="modal-header">
-                <button type="button" class="close" data-dismiss="modal" aria-label="Close"><span aria-hidden="true">&times;</span></button>
-                <h4 class="modal-title" id="myModalLabel">{{ troop_modal_title }}</h4>
-            </div>
-            <div class="modal-body">
-                <form method="POST" action="/register/troop/" id="troop_modal_form">
-                    {% csrf_token %}
-                    {{ troop_modal_form.as_p }}
-                </form>
-            <button type="button" class="btn btn-primary btn-lg" data-toggle="modal" data-target="#council_modal">
-                Council not listed? Add it
-            </button>
-            </div>
-            <div class="modal-footer">
-                <button type="button" class="btn btn-default" data-dismiss="modal">Close</button>
-                <button id="troop_submit_button" type="button" class="btn btn-primary" data-dismiss="modal">{{ troop_modal_title }}</button>
-            </div>
-        </div>
-    </div>
-</div>
