--- registration/templates/registration/register.html
@@ -1,8 +1,0 @@
-{% extends 'base.html' %}
-{% block content %}
-<a class='btn btn-danger' href="{% url 'register_scout' %}">I am a scout</a>
-<a class='btn btn-warning' href="{% url 'register_scoutmaster' %}">I am a scoutmaster</a>
-<a class='btn btn-success' href="{% url 'register_venture' %}">I am a venture</a>
-<a class='btn btn-info' href="{% url 'register_troopcontact' %}">I am a troop contact</a>
-<a class='btn btn-primary' href="{% url 'register_volunteer' %}">I am a volunteer</a>
-{% endblock %}
