--- registration/templates/registration/register_scout.html
@@ -1,18 +1,0 @@
-{% extends "base.html" %}
-{% block title %}
-	Register Scout
-{% endblock title %}
-
-{% block content %}
-    <form method="post" action=".">
-        <table>
-            {% csrf_token %}
-            {{ form }}
-            {{ formset }}
-            <tr>
-                <td> <input type="submit" value="Register" /> </td>
-            </tr>
-        </table>
-    </form>
-{% endblock %}
-
