--- registration/templates/registration/register_user.html
@@ -1,19 +1,0 @@
-{% extends "base.html" %}
-{% block title %}
-    {{ title }}
-{% endblock title %}
-
-{% block content %}
-    <form method="post" action=".">
-        <table>
-            {% csrf_token %}
-            {{ form }}
-            {{ formset }}
-            <tr>
-                <td> <input type="submit" value="Register" /> </td>
-            </tr>
-        </table>
-    </form>
-    {% include "registration/add_troop_modal.html" %}
-    {% include "registration/add_council_modal.html" %}
-{% endblock %}
