--- scout/admin.py
@@ -1,6 +1,0 @@
-from django.contrib import admin
-from scout.models import Scout
-
-# Register your models here.
-
-admin.site.register(Scout)
