--- scoutmaster/admin.py
@@ -1,5 +1,0 @@
-from django.contrib import admin
-from scoutmaster.models import Scoutmaster
-
-# Register your models here.
-admin.site.register(Scoutmaster)
