--- scoutmaster/templates/scoutmaster/view_troop.html
@@ -1,12 +1,0 @@
-{% extends 'base.html' %}
-{% block title %}Edit Troop Courses{% endblock title %}
-{% block header_text %}Edit Troop Courses{% endblock header_text %}
-
-{% block content %}
-    <ul>
-        {% for scout in scouts %}
-            <a href="{% url 'scoutmaster_view_classes' scout_id=scout.id %}"><li>{{ scout }}</li></a>
-        {% endfor %}
-    </ul>
-
-{% endblock content %}
