--- troop/admin.py
@@ -1,6 +1,0 @@
-from django.contrib import admin
-from troop.models import Troop, Council
-
-# Register your models here.
-admin.site.register(Troop)
-admin.site.register(Council)
