+++ utmbu/settings.py
@@ -58,13 +58,7 @@ INSTALLED_APPS = (
     'django.contrib.staticfiles',
     'social.apps.django_app.default',
     'mbu',
-    'course',
-    'scout',
-    'scoutmaster',
-    'registration',
-    'troop',
     'dev',
-    'mbu_users',
 )
 
 MIDDLEWARE_CLASSES = (
